+++ src/SillyDialect.cpp
@@ -5,21 +5,43 @@
 ///
 #include "SillyDialect.hpp"
 
+#include <llvm/ADT/TypeSwitch.h>
+#include <mlir/IR/Builders.h>
+#include <mlir/Dialect/LLVMIR/LLVMTypes.h>
+#include <mlir/IR/DialectImplementation.h>
+
+// Pull in generated op method bodies, adaptors, verify(), fold(), etc.
 #define GET_OP_CLASSES
 #include "SillyDialect.cpp.inc"
 
-// Include generated dialect definitions
-#include "SillyDialectDefs.cpp.inc"
+// Pull in generated type method bodies (parse, print, etc. if any)
+#define GET_TYPEDEF_CLASSES
+#include "SillyTypes.cpp.inc"
+
+using namespace mlir;
 
 namespace silly
 {
+
     void SillyDialect::initialize()
     {
+        // Register types
+        addTypes<
+#define GET_TYPEDEF_LIST
+#include "SillyTypes.cpp.inc"
+            >();
+
+        // Register operations
         addOperations<
 #define GET_OP_LIST
 #include "SillyDialect.cpp.inc"
             >();
+
+
+        //llvm::errs() << "All registrations complete. Testing type print...\n";
     }
 }    // namespace silly
+
+#include "SillyDialectDefs.cpp.inc"
 
 // vim: et ts=4 sw=4