+++ src/SillyPasses.hpp
@@ -1,6 +1,8 @@
+///
 /// @file    SillyPasses.hpp
 /// @author  Peeter Joot <peeterjoot@pm.me>
 /// @brief   Pass for MLIR lowering to LLVM-IR.
+///
 #pragma once
 
 #include <mlir/Pass/Pass.h>
@@ -16,9 +18,7 @@ namespace mlir
     {
         registerSillyToLLVMLoweringPass( pst );
     }
-}    // namespace mlir
 
-#define GEN_PASS_REGISTRATION
-#include "SillyPasses.hpp.inc"
+}    // namespace mlir
 
 // vim: et ts=4 sw=4