+++ src/SillyTypes.hpp
@@ -1,0 +1,16 @@
+///
+/// @file    SillyTypes.hpp
+/// @author  Peeter Joot <peeterjoot@pm.me>
+/// @brief   silly::varType definition
+///
+#pragma once
+
+#include <llvm/ADT/ArrayRef.h>
+#include <mlir/IR/Types.h>
+#include <mlir/IR/OpDefinition.h>
+
+// Generated type class definitions (Silly_VarType)
+#define GET_TYPEDEF_CLASSES
+#include "SillyTypes.hpp.inc"
+
+// vim: et ts=4 sw=4